requirements.txt
Flask==2.3.3
Flask-PyMongo==2.3.0
Flask-Login==0.6.2
pymongo==4.5.0
Werkzeug==2.3.7
python-dotenv==1.0.0
gunicorn==21.2.0
openai>=1.0.0
anthropic>=0.5.0
requests>=2.25.0
boto3>=1.28.0